refactor notebook snapshot menu into new file
